move config and context menus from WorkingSetView to SidebarView
