+++ src/htmlContent/main-view.html
@@ -41,12 +41,9 @@
     <!-- Main UI: horizontal set of sidebar, main content vertical stack, and vertical toolbar -->
     <div class="main-view">
         <div id="sidebar" class="sidebar panel quiet-scrollbars horz-resizable right-resizer collapsible" data-minsize="0" data-forceleft=".content">
-<!--
-            <div id="working-set-list-header">
-                <div class="working-set-splitview-btn btn-alt-quiet"></div>
-                <div class="working-set-option-btn btn-alt-quiet"></div>
-            <div>
--->
+            <div class="working-set-splitview-btn btn-alt-quiet"></div>
+            <div class="working-set-option-btn btn-alt-quiet"></div>
+            
             <div id="working-set-list-container">
                 
             </div>
