+++ src/project/WorkingSetView.js
@@ -39,7 +39,6 @@ define(function (require, exports, module) {
         CommandManager        = require("command/CommandManager"),
         Commands              = require("command/Commands"),
         Menus                 = require("command/Menus"),
-        DefaultMenus          = require("command/DefaultMenus"),
         FileViewController    = require("project/FileViewController"),
         ViewUtils             = require("utils/ViewUtils"),
         paneListTemplate      = require("text!htmlContent/working-set.html"),
@@ -56,22 +55,6 @@ define(function (require, exports, module) {
     var _views = [];
     
     /**
-     * Context Menu
-     * @private
-     * @type {Menu}
-     * 
-     */
-    var _workingset_cmenu;
-    
-    /**
-     * Context Menu
-     * @private
-     * @type {Menu}
-     * 
-     */
-    var _workingset_configuration_menu;
-    
-    /**
      * Constants for event.which values
      * @enum {number}
      */
@@ -86,26 +69,6 @@ define(function (require, exports, module) {
      */
     var _FILE_KEY = "file";
 
-    /* 
-     * Determines if context menus are registered
-     * @private
-     * @return {boolean} true if the menus are registered, false if not
-     */
-    function _areContextMenusRegistered() {
-        return _workingset_cmenu && _workingset_configuration_menu;
-    }
-    
-    /* 
-     * Registers context menus
-     * @private
-     */
-    function _registerContextMenus() {
-        if (!_areContextMenusRegistered()) {
-            _workingset_cmenu = Menus.getContextMenu(Menus.ContextMenuIds.WORKING_SET_CONTEXT_MENU);
-            _workingset_configuration_menu = Menus.getContextMenu(Menus.ContextMenuIds.WORKING_SET_CONFIG_MENU);
-        }
-    }
-    
     /** 
      * Updates the appearance of the list element based on the parameters provided.
      * @private
@@ -146,18 +109,9 @@ define(function (require, exports, module) {
         this.suppressSortRedraw = false;
         this.paneId = paneId;
         
-        this.updateOptionsButton();
         this.init();
     }
 
-    /*
-     * updates the visibility state of the gear button
-     */
-    WorkingSetView.prototype.updateOptionsButton = function () {
-        var visible = (MainViewManager.getActivePaneId() === this.paneId);
-        this.$el.find(".working-set-option-btn").toggle(visible);
-    };
-    
     /*
      * paneLayoutChange event listener
      * @private
@@ -171,7 +125,6 @@ define(function (require, exports, module) {
         }
         
         $titleEl.text(title);
-        this.updateOptionsButton();
     };
 
     /**
@@ -342,7 +295,6 @@ define(function (require, exports, module) {
         }
         this._adjustForScrollbars();
         this._fireSelectionChanged();
-        this.updateOptionsButton();
     };
     
     /**
@@ -793,7 +745,6 @@ define(function (require, exports, module) {
         
         this.$openFilesContainer = this.$el.find(".open-files-container");
         this.$workingSetListViewHeader = this.$el.find(".working-set-header");
-        this.$gearMenu = this.$el.find(".working-set-option-btn");
         
         this.$openFilesList = this.$el.find("ul");
         
@@ -818,46 +769,10 @@ define(function (require, exports, module) {
         // Disable horizontal scrolling until WebKit bug #99379 is fixed
         this.$openFilesContainer.css("overflow-x", "hidden");
 
-        this.installMenuHandlers();
-        
         this._redraw();
     };
 
 
-    /** 
-     * Installs the gear and context menu handlers
-     */
-    WorkingSetView.prototype.installMenuHandlers = function () {
-        var self = this;
-        
-        this.$openFilesContainer.on("contextmenu", function (e) {
-            _registerContextMenus();
-            _workingset_cmenu.open(e);
-        });
-
-        this.$gearMenu.on("click", function (e) {
-            var buttonOffset,
-                buttonHeight;
-
-            e.stopPropagation();
-
-            MainViewManager.setActivePaneId(self.paneId);
-            _registerContextMenus();
-            
-            if (_workingset_configuration_menu.isOpen()) {
-                _workingset_configuration_menu.close();
-            } else {
-                buttonOffset = $(this).offset();
-                buttonHeight = $(this).outerHeight();
-                _workingset_configuration_menu.open({
-                    pageX: buttonOffset.left,
-                    pageY: buttonOffset.top + buttonHeight
-                });
-            }
-        });
-            
-    };
-    
     /** 
      * Destroys the WorkingSetView DOM element and removes all event handlers
      */
