+++ src/styles/brackets.less
@@ -710,16 +710,17 @@ a, img {
     .sprite-icon(0, 0, 13px, 13px, "images/topcoat-settings-13.svg");
     background-position: center;
     opacity: 0.8;
+    z-index: 1;
 }
 
 .working-set-splitview-btn {
     position: absolute;
-    right: 4px;
+    right: 30px;
     top: 7px;
     padding: 4px 6px;
-    .sprite-icon(0, 0, 13px, 13px, url("images/split-view-icons.svg"));
+    .sprite-icon(0, 0, 13px, 13px, "images/split-view-icons.svg");
     background-position: center;
-    opacity: 0.8;
+    z-index: 1;
 }
 
 #sidebar {
